Revert "Replace snake and ludo tokens with chess bishops"
